medical-term-importer/meilisearch/python/requirements.txt
meilisearch==0.31.6
python-dotenv==1.0.0
psutil==5.9.8
pydantic==2.9.2
